Change signature of outshine-define-key-with-fallback
Pass the command to call as a quoted symbol instead of an unquoted function
call, matching the signature of `define-key'.

Also remove the unhygenic variable capture of "arg" in the macro, since the
command is now being called interactively.

Note: this is a breaking change if the macro was used to define personal
keybindings outside of outshine.el