more fully override imenu configuration when using outshine's imenu
When using outshine's imenu functionality (from either `outshine-imenu` or
`outshine-imenu-with-navi-regexp`), the variables that configure imenu behavior
are locally overwritten in a `let*` form. However, it's important to also restore
the default setting of `imenu-create-index-function` (which should be
`imenu-default-create-index-function`).

I noticed this bug when editing `R` files in `spacemacs`: the `lsp`
configuration involves changing the value of `imenu-create-index-function`,
which makes changes to the other variables inconsequential, and then outshine's
attempt to change the way imenu works by binding `imenu-generic-expression` has
no effect.